requirements.txt
ultralytics
torch
torchvision
opencv-python
matplotlib
numpy
PyYAML
pandas
ipython
tqdm
seaborn
psutil
